test(e2e): move the shared rollout helpers out of the cleanup test
