+++ test/e2e/set/scyllacluster/scyllacluster_cleanup.go
@@ -4,22 +4,18 @@ package scyllacluster
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	g "github.com/onsi/ginkgo/v2"
 	o "github.com/onsi/gomega"
 	scyllav1 "github.com/scylladb/scylla-operator/pkg/api/scylla/v1"
-	"github.com/scylladb/scylla-operator/pkg/controllerhelpers"
 	"github.com/scylladb/scylla-operator/pkg/helpers"
 	"github.com/scylladb/scylla-operator/pkg/naming"
 	"github.com/scylladb/scylla-operator/test/e2e/framework"
 	"github.com/scylladb/scylla-operator/test/e2e/utils"
-	scyllaclusterverification "github.com/scylladb/scylla-operator/test/e2e/utils/verification/scyllacluster"
 	batchv1 "k8s.io/api/batch/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/tools/cache"
 )
@@ -41,8 +37,7 @@ var _ = g.Describe("ScyllaCluster", framework.SuiteParallel, framework.SuitePara
 		err := jobObserver.Start(ctx)
 		o.Expect(err).NotTo(o.HaveOccurred())
 
-		sc, err := createClusterAndWaitForRollout(ctx, f, e.initialMembers)
-		o.Expect(err).NotTo(o.HaveOccurred())
+		sc := createClusterAndWaitForRollout(ctx, f, e.initialMembers)
 
 		verifyCleanupJobsCreatedEventually(ctx, f, sc, &jobObserver)
 
@@ -53,8 +48,7 @@ var _ = g.Describe("ScyllaCluster", framework.SuiteParallel, framework.SuitePara
 		err = jobObserver.Start(ctx)
 		o.Expect(err).NotTo(o.HaveOccurred())
 
-		sc, err = scaleClusterAndWaitForRollout(ctx, f, sc, e.targetMembers)
-		o.Expect(err).NotTo(o.HaveOccurred())
+		sc = scaleClusterAndWaitForRollout(ctx, f, sc, e.targetMembers)
 
 		verifyCleanupJobsCreatedEventually(ctx, f, sc, &jobObserver)
 
@@ -77,8 +71,7 @@ var _ = g.Describe("ScyllaCluster", framework.SuiteParallel, framework.SuitePara
 		err := jobObserver.Start(ctx)
 		o.Expect(err).NotTo(o.HaveOccurred())
 
-		sc, err := createClusterAndWaitForRollout(ctx, f, 3)
-		o.Expect(err).NotTo(o.HaveOccurred())
+		sc := createClusterAndWaitForRollout(ctx, f, 3)
 
 		verifyCleanupJobsCreatedEventually(ctx, f, sc, &jobObserver)
 
@@ -133,60 +126,4 @@ func verifyCleanupJobsCreatedEventually(
 
 		eo.Expect(cleanedUpNodes).To(o.ConsistOf(memberServiceNames))
 	}).WithTimeout(30 * time.Second).WithPolling(1 * time.Second).Should(o.Succeed())
-}
-
-// createClusterAndWaitForRollout creates a ScyllaCluster with the specified number of members and waits for rollout.
-func createClusterAndWaitForRollout(
-	ctx context.Context,
-	f *framework.Framework,
-	members int32,
-) (*scyllav1.ScyllaCluster, error) {
-	sc := f.GetDefaultScyllaCluster()
-	sc.Spec.Datacenter.Racks[0].Members = members
-
-	framework.By("Creating a %d node ScyllaCluster", members)
-	sc, err := f.ScyllaClient().ScyllaV1().ScyllaClusters(f.Namespace()).Create(ctx, sc, metav1.CreateOptions{})
-	o.Expect(err).NotTo(o.HaveOccurred())
-
-	framework.By("Waiting for the ScyllaCluster to roll out (RV=%s)", sc.ResourceVersion)
-	waitCtx, waitCtxCancel := utils.ContextForRollout(ctx, sc)
-	defer waitCtxCancel()
-	sc, err = controllerhelpers.WaitForScyllaClusterState(waitCtx, f.ScyllaClient().ScyllaV1().ScyllaClusters(sc.Namespace), sc.Name, controllerhelpers.WaitForStateOptions{}, utils.IsScyllaClusterRolledOut)
-	o.Expect(err).NotTo(o.HaveOccurred())
-
-	scyllaclusterverification.Verify(ctx, f.KubeClient(), f.ScyllaClient(), sc)
-
-	return sc, nil
-}
-
-// scaleClusterAndWaitForRollout scales the cluster to the given number of members and waits for rollout.
-func scaleClusterAndWaitForRollout(
-	ctx context.Context,
-	f *framework.Framework,
-	sc *scyllav1.ScyllaCluster,
-	members int32,
-) (*scyllav1.ScyllaCluster, error) {
-	patchData := []byte(fmt.Sprintf(`[{"op": "replace", "path": "/spec/datacenter/racks/0/members", "value": %d}]`, members))
-
-	framework.By("Scaling the ScyllaCluster to %d members", members)
-	sc, err := f.ScyllaClient().ScyllaV1().ScyllaClusters(f.Namespace()).Patch(
-		ctx,
-		sc.Name,
-		types.JSONPatchType,
-		patchData,
-		metav1.PatchOptions{},
-	)
-	o.Expect(err).NotTo(o.HaveOccurred())
-	o.Expect(sc.Spec.Datacenter.Racks).To(o.HaveLen(1))
-	o.Expect(sc.Spec.Datacenter.Racks[0].Members).To(o.BeEquivalentTo(members))
-
-	framework.By("Waiting for the ScyllaCluster to roll out (RV=%s)", sc.ResourceVersion)
-	waitCtx, waitCtxCancel := utils.ContextForRollout(ctx, sc)
-	defer waitCtxCancel()
-	sc, err = controllerhelpers.WaitForScyllaClusterState(waitCtx, f.ScyllaClient().ScyllaV1().ScyllaClusters(sc.Namespace), sc.Name, controllerhelpers.WaitForStateOptions{}, utils.IsScyllaClusterRolledOut)
-	o.Expect(err).NotTo(o.HaveOccurred())
-
-	scyllaclusterverification.Verify(ctx, f.KubeClient(), f.ScyllaClient(), sc)
-
-	return sc, nil
 }
